It's Alive! (see description)
Importing combatlogs registers players and performances in the database. Preview of the data in the GUI is extremely limited right now, limited to 10/20 entries, and its not even showing all the properties of the entities yet.

So baseline functionality is there, and now its time to get started on making a usable application GUI. 
Gotta check if theres anything for wpf xaml to make it easier to manage.