randomly moved around the comboboxes so they would always show
There is a really strange issue where thebackground of the window/grid or whatever would force itself to be in front of the comboboxes? So i just moved them around and back into the same spot where they were before, and now the issue magically disappeared. I do not understand!