Introduce ProgressBar for Import (UX Improvement Issue Done)
Importing can now be done entirely asynchronously.
Upside: well, the ability to show the progress of the import in the GUI.
Downside: It is noticably slower when fetching the encounter metadata. Might have to consider keeping that part synchronously to just get it done with, as it is essentially a single step where no updates happen??