Increase FileStream buffer size
In a short little benchmark, increasing the buffer size by a factor of 4 (from 4096 to 16384), it resulted in a 25% increase in performance when reading an entire file.
This will overall be a tiny amount of the actual time it takes to parse through a log, but its essentially a free performance gain, so there is no reason not to do it.